Symlink chatyuk/app to node_modules/app
We do this so that we can easily require all our
app related source files from a common root
instead using confusing relative paths.

This is the method recommended by the browserify handbook
https://github.com/substack/browserify-handbook#avoiding-

It feels a bit ghetto to add a symlink to the repo like this
but given all other solutions require using an additional
npm package (alisify, browserify-shim) this seemed like the
most lightweight way to achieve it.